Hook up submitting a call mutation. Closes #110.

+++ src/backend/data/models/bsd-email.js
@@ -1,7 +1,7 @@
 export default function(sequelize, DataTypes) {
   return sequelize.define('BSDEmail', {
     id: {
-      type: DataTypes.BIGINT,
+      type: DataTypes.INTEGER,
       field: 'cons_email_id',
       primaryKey: true
     },
